added build script to generate a distributable version in dist/
Builds are made with webpack, leaving jQuery as an external dependency
to be provided by the user of the package.

The package internally depends on lodash. Since it isn't as widely used
as jQuery, That dependency is resolved adding a custom lodash build
(the small subset being used by the package) to the distributable version.